+++ simba-core/src/main/java/org/simbasecurity/core/domain/SSOTokenMappingEntity.java
@@ -69,7 +69,7 @@ public class SSOTokenMappingEntity implements SSOTokenMapping {
     }
 
     private long getMaxLoginElapsedTime() {
-        Integer maxElapsedTime = getConfigurationService().getValue(MAX_LOGIN_ELAPSED_TIME);
+        Long maxElapsedTime = getConfigurationService().getValue(MAX_LOGIN_ELAPSED_TIME);
         return Duration.of(maxElapsedTime, MAX_LOGIN_ELAPSED_TIME.getChronoUnit()).toMillis();
     }
